requirements.txt
# Upgrade pip/setuptools/wheel in render.yaml before install
fastapi>=0.104.0,<0.115.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
joblib>=1.3.0
python-multipart>=0.0.6
ultralytics>=8.1.0
pillow>=10.0.0
